make term page config

+++ src/main/java/ir/golestan/controller/TermController.java
@@ -32,10 +32,10 @@ public class TermController {
         return "home";
     }
     @RequestMapping(value = "/maketerm" ,method = RequestMethod.POST)
-    public String makeTerm(@RequestParam("code")Long code,@RequestParam("year")int year,@RequestParam("season")int season,Model model){
+    public String makeTerm(@RequestParam("code")int code,@RequestParam("year")int year,@RequestParam("season")int season,Model model){
         if(httpSession.getAttribute("username")!=null&&((String)httpSession.getAttribute("rule")).compareTo("admin")==0){
            Term term = new Term();
-            term.setCode(code);
+            term.setCode((long) code);
             term.setSeason(season);
             term.setYear(year);
             termService.create(term);
